requirements.txt
torch
torchvision
torchaudio
timm
albumentations
opencv-python
numpy
pillow
tqdm
hydra-core
omegaconf
streamlit
google-generativeai